Add workarround for "Docker Desktop" for Linux
With "Docker Desktop", docker is run as a user.
Inside the container, all processes operate as root, regardless of the user/group ID specified. In directories that are mapped into the containers, the folders and files belong to the user who is currently using "Docker Desktop". 
This is the same behavior as on Windows or Mac.
In order not to get permission errors when executing commands inside the container, the CLI of the container is called as root in this case.
For users who continue to use Docker without "Docker Desktop" on Linux, the user in the CLI is still used.